LayoutTests/fast/viewport/viewport-133-expected.txt

CONSOLE MESSAGE: line 3: Viewport argument value "10;" for key "minimum-scale" was truncated to its numeric prefix.
ALERT: viewport size 980x1078 scale 10 with limits [10, 10] and userScalable true
